Refactoring: Move signature verification stuff to consumer subpacket
