+++ pgpainless-core/src/main/java/org/pgpainless/key/KeyRingValidator.java
@@ -19,8 +19,8 @@ import org.pgpainless.algorithm.SignatureType;
 import org.pgpainless.exception.SignatureValidationException;
 import org.pgpainless.implementation.ImplementationFactory;
 import org.pgpainless.policy.Policy;
-import org.pgpainless.signature.SignatureCreationDateComparator;
-import org.pgpainless.signature.SignatureVerifier;
+import org.pgpainless.signature.consumer.SignatureCreationDateComparator;
+import org.pgpainless.signature.consumer.SignatureVerifier;
 import org.pgpainless.util.CollectionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
