+++ pgpainless-core/src/main/java/org/pgpainless/signature/consumer/SignatureVerifier.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.pgpainless.signature;
+package org.pgpainless.signature.consumer;
 
 import java.io.IOException;
 import java.io.InputStream;
