+++ pgpainless-core/src/test/java/org/pgpainless/signature/BindingSignatureSubpacketsTest.java
@@ -22,6 +22,7 @@ import org.pgpainless.PGPainless;
 import org.pgpainless.exception.SignatureValidationException;
 import org.pgpainless.implementation.ImplementationFactory;
 import org.pgpainless.policy.Policy;
+import org.pgpainless.signature.consumer.CertificateValidator;
 
 /**
  * Explores how subpackets on binding sigs are handled.
