+++ pgpainless-core/src/test/java/org/pgpainless/signature/SignatureOverUserAttributesTest.java
@@ -28,6 +28,7 @@ import org.pgpainless.implementation.ImplementationFactory;
 import org.pgpainless.key.TestKeys;
 import org.pgpainless.key.protection.SecretKeyRingProtector;
 import org.pgpainless.key.protection.UnlockSecretKey;
+import org.pgpainless.signature.consumer.SignatureVerifier;
 
 public class SignatureOverUserAttributesTest {
